projects/ninja-dashboard/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
aiofiles>=24.1.0
psycopg2-binary>=2.9.9
celery[redis]>=5.3
yt-dlp>=2024.12.0
requests>=2.31.0
keyring>=25.0.0
google-genai>=1.0.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.2.0
